Move scrap paper library to subdirectory
