+++ lib/scrap-paper/generate.js
@@ -1,7 +1,7 @@
 const jsCrc = require('js-crc');
 
-const sqldb = require('../prairielib/lib/sql-db');
-const sqlLoader = require('../prairielib/lib/sql-loader');
+const sqldb = require('../../prairielib/lib/sql-db');
+const sqlLoader = require('../../prairielib/lib/sql-loader');
 const sql = sqlLoader.loadSqlEquiv(__filename);
 
 const bitgener = require('bitgener');
