+++ lib/scrap-paper/generate.sql
