+++ lib/scrap-paper/read.js
@@ -1,8 +1,9 @@
 const pdfjsLib = require('pdfjs-dist/legacy/build/pdf.js'); // legacy means node support
 
 /**
- * Load metadata about pdf
+ * Load metadata about PDF.
  * https://github.com/mozilla/pdf.js/blob/master/examples/node/getinfo.js
+ *
  * @param {Buffer} pdfBuffer
  * @returns {object} pdf.js wrapper
  */
